refactor: change router folder name
